Fix kernel and spv file

+++ stress_tests/test_commands_overloading/kernels/test_commands_overloading.cl
@@ -7,6 +7,6 @@
  */
 
 kernel void test_device_memory1(__global uint *out) {
-    int  tid = get_global_id(0);
+    size_t tid = get_global_id(0);
     out[tid] = out[tid] + tid;
 }
